feat: memory tool trio wired into builtin_tools, default-off (#333)
M2 batch 3: memory tool trio (memory_append/memory_search/memory_list) wired into builtin_tools behind a default-off env gate, plus the memory API completion (append returns its id, RecallHit carries text) and an rpc spawn retry for transient ETXTBSY that was flaking CI.

+++ crates/memory/src/inject.rs
@@ -227,7 +227,7 @@ fn overlaps_enough(last: &HashSet<String>, next: &HashSet<String>) -> bool {
     let shared = last.intersection(next).count();
     let baseline = last.len().max(next.len());
     // u64 math: usize overflow is theoretical here but the widening is free.
-    (shared as u64) * 100 >= u64::from(OVERLAP_MIN_PCT) * baseline as u64
+    (shared as u64) * 100 >= OVERLAP_MIN_PCT * baseline as u64
 }
 
 /// Collapse every whitespace run to one space and trim the ends — the one
